// src/types/data.ts
type RecordType = {
  title: string;
  contractorName: string;
  nip: string;
  paymentAccount: string;
  paymentSwift: string;
  grossAmount: string;
  currency: string;
  invoiceDateCreated: string;
  invoiceDateReceived: string;
  invoiceDateSold: string;
  payment: string;
  commentPayment: string;
}

export type { RecordType }